Implemented Twitter post on submission, removed some warnings
To gradually start using CFP exchange I’ve removed some warnings statig
to not use the system. It still needs (a lot) of love, but the basic
happy flow works.

Whenever a CFP is submitted, it is now posted to the Twitter feed.
Also added configuration for this and updated the README accordingly