Made all details of submitting a CFP work
Fix #18. Also does a lot more. Unfortunately some user account models have snuck in here, so the migration is poisoned.

Also fixes some small UI problems, makes sure an image is always shown for an event and adds some extra details to the CFP detail screen. A Tweet will not be sent when not in production. I think that covers it all...