Added ability to add event to your calendar
As per #30 there is now a button to add the event to your calendar if the dates are known. Implemented attempt to retrieve the timezone for a location to support this.

Also added grayscale when a CFP has ended but you still land on the page. Also, disabled the ability to report an issue for a passed CFP.